version updated

+++ build/controllers/user.controller.js
@@ -1,44 +1,176 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getUserDetails = exports.updateUserDetails = exports.deleteUser = exports.loginUser = exports.registerUser = void 0;
+exports.getPaymentHistory = exports.getReferredUsers = exports.logoutUser = exports.getUserDetails = exports.updateUserDetails = exports.deleteUser = exports.loginUser = exports.registerUser = void 0;
 const security_service_1 = require("../services/security.service");
 const user_1 = require("../data/user");
 const user_2 = require("../modals/user");
+const user_details_1 = require("../modals/user-details");
 const userid_service_1 = require("../services/userid.service");
 const active_user_1 = require("../modals/active-user");
 const jwt = require("jsonwebtoken");
 const constant_1 = require("../data/constant");
+const user_level_1 = require("../modals/user-level");
+const payment_history_1 = require("../modals/payment-history");
 const registerUser = async (req, res) => {
+    const { email, password, firstname, lastname, confirmpassword, mobilenumber, } = req.body;
+    if (!email) {
+        return res.status(400).send({
+            message: "Email is required",
+            statusCode: 400,
+        });
+    }
+    else if (!password) {
+        return res.status(400).send({
+            message: "Password is required",
+            statusCode: 400,
+        });
+    }
+    else if (!firstname) {
+        return res.status(400).send({
+            message: "firstname is required",
+            statusCode: 400,
+        });
+    }
+    else if (!lastname) {
+        return res.status(400).send({
+            message: "lastname is required",
+            statusCode: 400,
+        });
+    }
+    else if (!mobilenumber) {
+        return res.status(400).send({
+            message: "mobilenumber is required",
+            statusCode: 400,
+        });
+    }
+    else if (!confirmpassword) {
+        return res.status(400).send({
+            message: "confirmpassword is required",
+            statusCode: 400,
+        });
+    }
     if (req.body.password !== req.body.confirmpassword) {
         return res.status(400).send({
             message: "Password and confirm password do not match",
             statusCode: 400,
         });
     }
-    const hashedPassword = await (0, security_service_1.hashPassword)(req.body.password);
-    const userid = await (0, userid_service_1.generateUserId)();
-    const newUser = new user_2.Users({
-        firstname: req.body.firstname,
-        middlename: req.body.middlename,
-        lastname: req.body.lastname,
-        email: req.body.email,
-        password: hashedPassword,
-        userid: userid
-    });
     try {
-        const existingUser = await user_2.Users.findOne({ email: req.body.email });
-        if (existingUser) {
-            return res.status(400).send({
-                message: "Email already exists",
-                statusCode: 400,
+        const referDetails = req.body?.referralcode;
+        const userid = await (0, userid_service_1.generateUserId)();
+        let side = 3;
+        let referbyuserid = 0;
+        if (referDetails) {
+            side = +referDetails.substr(referDetails.length - 1);
+            referbyuserid =
+                (+referDetails.substring(0, referDetails.length - 1) - 99) / 10;
+            const newUser = new user_2.Users({
+                firstname: req.body.firstname,
+                lastname: req.body.lastname,
+                email: req.body.email,
+                mobilenumber: req.body.mobilenumber,
+                referbyuserid: +referbyuserid,
+                referralcode: 10 * userid + 99,
+                accounttypeside: +side,
+                password: req.body.password,
+                userid: userid,
             });
+            await user_level_1.UserLevels.findOneAndUpdate({ userid: +referbyuserid }, { $addToSet: { level1: {
+                        userid: userid,
+                        accounttypeside: +side,
+                        status: 2
+                    } } }, { new: true, upsert: true });
+            const level1User = await user_2.Users.findOne({ userid: +referbyuserid });
+            if (level1User?.referbyuserid) {
+                await user_level_1.UserLevels.findOneAndUpdate({ userid: +level1User?.referbyuserid }, { $addToSet: { level2: {
+                            userid: userid,
+                            accounttypeside: +side,
+                            status: 2
+                        } } }, { new: true, upsert: true });
+            }
+            const level2User = await user_2.Users.findOne({
+                userid: +level1User?.referbyuserid,
+            });
+            if (level2User?.referbyuserid) {
+                await user_level_1.UserLevels.findOneAndUpdate({ userid: +level2User?.referbyuserid }, { $addToSet: { level3: {
+                            userid: userid,
+                            accounttypeside: +side,
+                            status: 2
+                        } } }, { new: true, upsert: true });
+                if (level2User?.referbyuserid) {
+                    const level3User = await user_2.Users.findOne({
+                        userid: +level2User?.referbyuserid,
+                    });
+                    if (level3User?.referbyuserid) {
+                        await user_level_1.UserLevels.findOneAndUpdate({ userid: +level3User?.referbyuserid }, { $addToSet: { level4: {
+                                    userid: userid,
+                                    accounttypeside: +side,
+                                    status: 2
+                                } } }, { new: true, upsert: true });
+                        const level4User = await user_2.Users.findOne({
+                            userid: +level3User?.referbyuserid,
+                        });
+                        if (level4User?.referbyuserid) {
+                            await user_level_1.UserLevels.findOneAndUpdate({ userid: +level4User?.referbyuserid }, { $addToSet: { level5: {
+                                        userid: userid,
+                                        accounttypeside: +side,
+                                        status: 2
+                                    } } }, { new: true, upsert: true });
+                            const level5User = await user_2.Users.findOne({
+                                userid: +level4User?.referbyuserid,
+                            });
+                            if (level5User?.referbyuserid) {
+                                await user_level_1.UserLevels.findOneAndUpdate({ userid: +level5User?.referbyuserid }, { $addToSet: { level6: {
+                                            userid: userid,
+                                            accounttypeside: +side,
+                                            status: 2
+                                        } } }, { new: true, upsert: true });
+                            }
+                        }
+                    }
+                }
+            }
+            await user_level_1.UserLevels.findOneAndUpdate({ userid: userid }, {
+                $set: {
+                    referbyuserid: +referbyuserid,
+                    level1: [],
+                    level2: [],
+                    level3: [],
+                    level4: [],
+                    level5: [],
+                    level6: [],
+                },
+            }, { new: true, upsert: true });
+            await newUser.save();
         }
-        await newUser.save();
-        const recieverDetails = {
-            name: newUser.firstname,
-            email: newUser.email
-        };
-        res.status(200).send(user_1.registerUserResponse);
+        else {
+            const userid = await (0, userid_service_1.generateUserId)();
+            const newUser = new user_2.Users({
+                firstname: req.body.firstname,
+                lastname: req.body.lastname,
+                email: req.body.email,
+                mobilenumber: req.body.mobilenumber,
+                referbyuserid: +referbyuserid,
+                referralcode: 10 * userid + 99,
+                accounttypeside: +side,
+                password: req.body.password,
+                userid: userid,
+            });
+            await newUser.save();
+            await user_level_1.UserLevels.findOneAndUpdate({ userid: userid }, {
+                $set: {
+                    referbyuserid: +referbyuserid,
+                    level1: [],
+                    level2: [],
+                    level3: [],
+                    level4: [],
+                    level5: [],
+                    level6: [],
+                },
+            }, { new: true, upsert: true });
+        }
+        const registerResponse = { ...user_1.registerUserResponse, userid };
+        res.status(200).send(registerResponse);
     }
     catch (err) {
         res.status(500).send(err);
@@ -47,10 +179,10 @@ const registerUser = async (req, res) => {
 exports.registerUser = registerUser;
 const loginUser = async (req, res) => {
     try {
-        const { email, password } = req.body;
-        if (!email) {
+        const { userid, password } = req.body;
+        if (!userid) {
             return res.status(400).send({
-                message: "Email is required",
+                message: "userid is required",
                 statusCode: 400,
             });
         }
@@ -60,34 +192,36 @@ const loginUser = async (req, res) => {
                 statusCode: 400,
             });
         }
-        const existingUser = await user_2.Users.findOne({ email });
+        const existingUser = await user_2.Users.findOne({ userid });
         if (!existingUser) {
             return res.status(400).send({
                 message: "Invalid credentials",
                 statusCode: 400,
             });
         }
-        const passwordMatches = await (0, security_service_1.verifyPassword)(password, existingUser.password);
-        if (!passwordMatches) {
+        if (!(password == existingUser.password)) {
             return res.status(400).send({
                 message: "Invalid credentials",
                 statusCode: 400,
             });
         }
         const newActiveUser = new active_user_1.ActiveUsers({
-            email: req.body.email,
-            userid: existingUser.userid
+            userid: existingUser.userid,
         });
         await newActiveUser.save();
-        const token = jwt.sign({ userid: existingUser.userid, email: email }, constant_1.JWT_SECRET_KEY, { expiresIn: '1h' });
+        const token = jwt.sign({ userid: existingUser.userid }, constant_1.JWT_SECRET_KEY, {
+            expiresIn: "24hr",
+        });
         const encrypteduserId = await (0, security_service_1.userIdEncryption)(existingUser.userid.toString());
         res.status(200).send({
             message: "Login successful",
             statusCode: 200,
             data: {
                 accountid: encrypteduserId,
-                bearertoken: token
-            }
+                bearertoken: token,
+                status: existingUser.status,
+                referralcode: existingUser.referralcode,
+            },
         });
     }
     catch (error) {
@@ -100,6 +234,13 @@ const loginUser = async (req, res) => {
 };
 exports.loginUser = loginUser;
 const deleteUser = async (req, res) => {
+    const { email } = req.body;
+    if (!email) {
+        return res.status(400).send({
+            message: "Email is required",
+            statusCode: 400,
+        });
+    }
     try {
         const email = req.body.email;
         await user_2.Users.deleteOne({ email: email });
@@ -111,16 +252,39 @@ const deleteUser = async (req, res) => {
 };
 exports.deleteUser = deleteUser;
 const updateUserDetails = async (req, res) => {
+    const accountid = req.headers.accountid;
+    if (!accountid) {
+        return res.status(400).send({
+            message: "Account id is required",
+            statusCode: 400,
+        });
+    }
+    const encrypteduserId = await (0, security_service_1.userIdDecryption)(accountid.toString());
     try {
-        const email = req.params.email;
-        const userDetail = await user_2.Users.findByIdAndUpdate({ email: email }, {
-            $set: {
-                firstname: req.body.firstname,
-                middlename: req.body.middlename,
-                lastname: req.body.lastname,
-                password: req.body.password,
-            },
-        }, { new: true });
+        let userDetail;
+        if (req.body.password) {
+            userDetail = await user_2.Users.findOneAndUpdate({ userid: +encrypteduserId }, {
+                $set: {
+                    password: req.body.password,
+                },
+            }, { new: true });
+        }
+        else {
+            userDetail = await user_2.Users.findOneAndUpdate({ userid: +encrypteduserId }, {
+                $set: {
+                    firstname: req.body.firstname,
+                    lastname: req.body.lastname,
+                    mobilenumber: req.body.mobilenumber,
+                },
+            }, { new: true });
+        }
+        if (req.body.location) {
+            await user_details_1.UserAccountDetails.findOneAndUpdate({ userid: userDetail?.userid }, {
+                $set: {
+                    location: req.body.location,
+                },
+            }, { new: true, upsert: true });
+        }
         const response = {
             data: userDetail,
             message: "success",
@@ -134,12 +298,20 @@ const updateUserDetails = async (req, res) => {
 };
 exports.updateUserDetails = updateUserDetails;
 const getUserDetails = async (req, res) => {
+    const accountid = req.headers.accountid;
+    if (!accountid) {
+        return res.status(400).send({
+            message: "Account id is required",
+            statusCode: 400,
+        });
+    }
+    const encrypteduserId = await (0, security_service_1.userIdDecryption)(accountid.toString());
     try {
-        const user = await user_2.Users.findById({ email: req.params.email });
+        const user = await user_2.Users.findOne({ userid: +encrypteduserId });
         const responseBody = {
             data: user,
             message: "success",
-            statusCode: 200
+            statusCode: 200,
         };
         res.status(200).send(responseBody);
     }
@@ -148,4 +320,68 @@ const getUserDetails = async (req, res) => {
     }
 };
 exports.getUserDetails = getUserDetails;
+const logoutUser = async (req, res) => {
+    const accountid = req.headers.accountid;
+    if (!accountid) {
+        return res.status(400).send({
+            message: "Account id is required",
+            statusCode: 400,
+        });
+    }
+    const encrypteduserId = await (0, security_service_1.userIdDecryption)(accountid.toString());
+    try {
+        await active_user_1.ActiveUsers.deleteMany({ userid: +encrypteduserId });
+        res.status(200).send(user_1.logoutUserResponse);
+    }
+    catch (error) {
+        res.status(500).send(error);
+    }
+};
+exports.logoutUser = logoutUser;
+const getReferredUsers = async (req, res) => {
+    const accountid = req.headers.accountid;
+    if (!accountid) {
+        return res.status(400).send({
+            message: "Account id is required",
+            statusCode: 400,
+        });
+    }
+    const encrypteduserId = await (0, security_service_1.userIdDecryption)(accountid.toString());
+    try {
+        const user = await user_level_1.UserLevels.find({ userid: encrypteduserId });
+        const responseBody = {
+            data: user ? user[0] : {},
+            message: "success",
+            statusCode: 200,
+        };
+        res.status(200).send(responseBody);
+    }
+    catch (error) {
+        res.status(500).send(error);
+    }
+};
+exports.getReferredUsers = getReferredUsers;
+const getPaymentHistory = async (req, res) => {
+    const accountid = req.headers.accountid;
+    if (!accountid) {
+        return res.status(400).send({
+            message: "Account id is required",
+            statusCode: 400,
+        });
+    }
+    const encrypteduserId = await (0, security_service_1.userIdDecryption)(accountid.toString());
+    try {
+        const user = await payment_history_1.PaymentHistoryUser.find({ userid: +encrypteduserId });
+        const responseBody = {
+            data: user,
+            message: "success",
+            statusCode: 200,
+        };
+        res.status(200).send(responseBody);
+    }
+    catch (error) {
+        res.status(500).send(error);
+    }
+};
+exports.getPaymentHistory = getPaymentHistory;
 //# sourceMappingURL=user.controller.js.map
